Target-Swapper, Triple Kick, Exit-Battle Moves, Divide by Zero Run Away,  Powder-free Grass, Re-effectiveness
Effects that affect move targets (Snatch, Magic Coat, Magic Bouce) are now generalized instead of stupidly being stupid hardcoded in the getTarget method
Included a TargetSwapperEffect interface for these effects
Magic Coat and Magic Bounce are generalized inside the generator

Switched Triple Kick to just be a regular multistrike move that hits exactly three times --  no longer performs an accuracy check at each hit with increasing power, power is just base 20 now
Fixed a bug with multistrike moves where the minimum number of hits is always essentially 2

Fixed a few problems with exit-battle moves like Roar and such
No longer fails in a wild battle -- when used by a wild pokemon, will switch the user's Pokemon to a random one, when used by the user, will end the battle if the wild pokemon is not at a higher level (fails if it is)

Fixed a bug with the game crashing when trying to run from a wild Pokemon whose Speed stat is less than 4

Powder moves no longer effect Grass-type Pokemon

Redid effectiveness in the battle system
No longer hard-coding blocking abilities in all the relevant places
effective-check now takes place in apply, instead of apply damage -- always returns true for self-target moves and field moves, checks zero type advantage for non-status moves and thunder wave, and checks ability and type blocking moves for everything

+++ src/battle/Battle.java
@@ -404,7 +404,7 @@ public class Battle
 		int pSpeed = Stat.getStat(Stat.SPEED, plyr, opp, this);
 		int oSpeed = Stat.getStat(Stat.SPEED, opp, plyr, this);
 		
-		int val = (pSpeed*32)/(oSpeed/4) + 30*escapeAttempts;
+		int val = (int)((pSpeed*32.0)/(oSpeed/4.0) + 30.0*escapeAttempts);
 		if (Math.random()*256 < val 
 				|| plyr.getAbility() instanceof DefiniteEscape 
 				|| plyr.getHeldItem(this) instanceof DefiniteEscape)
